pyproject.toml
[tool.poetry]
name = "gateway"
version = "0.1.0"
description = ""
authors = ["jhnnsrs <jhnnsrs@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
dask-gateway = "^2023.9.0"
dask-gateway-server = {extras = ["local"], version = "^2023.9.0"}
aiohttp = "^3.9.1"
pydantic = "<2"
pyjwt = "^2.8.0"
omegaconf = "^2.3.0"
bokeh = "^3.3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
